requirements.txt
numpy==2.2.3
pandas==2.3.3
scipy==1.17.0
scikit-learn==1.8.0
matplotlib==3.10.0
seaborn==0.13.2
torch==2.6.0+cu124
torchvision==0.21.0+cu124
Pillow==12.2.0
